requirements.txt
streamlit>=1.20.0
google-genai>=0.4.0
python-dotenv>=1.0.0
